Changed gulp tasks names: removed dev- prefix, changed word order for create-package task.
