+++ dev/tasks/dev/tasks.js
@@ -4,28 +4,28 @@
 
 const gulp = require( 'gulp' );
 const minimist = require( 'minimist' );
-const statusTask = require( './tasks/dev-status' );
-const initTask = require( './tasks/dev-init' );
-const installTask = require( './tasks/dev-install' );
-const pluginCreateTask = require( './tasks/dev-package-create' );
-const updateTask = require( './tasks/dev-update' );
-const relinkTask = require( './tasks/dev-relink' );
+const statusTask = require( './tasks/status' );
+const initTask = require( './tasks/init' );
+const installTask = require( './tasks/install' );
+const pluginCreateTask = require( './tasks/create-package' );
+const updateTask = require( './tasks/update' );
+const relinkTask = require( './tasks/relink' );
 
 module.exports = ( config ) => {
 	const ckeditor5Path = process.cwd();
 	const packageJSON = require( '../../../package.json' );
 
-	gulp.task( 'dev-init', () => {
+	gulp.task( 'init', () => {
 		initTask( installTask, ckeditor5Path, packageJSON, config.WORKSPACE_DIR, console.log );
 	} );
 
-	gulp.task( 'dev-package-create', ( done ) => {
+	gulp.task( 'create-package', ( done ) => {
 		pluginCreateTask( ckeditor5Path, config.WORKSPACE_DIR, console.log )
 			.then( done )
 			.catch( ( error )  => done( error ) );
 	} );
 
-	gulp.task( 'dev-update', () => {
+	gulp.task( 'update', () => {
 		const options = minimist( process.argv.slice( 2 ), {
 			boolean: [ 'npm-update' ],
 			default: {
@@ -36,15 +36,15 @@ module.exports = ( config ) => {
 		updateTask( ckeditor5Path, packageJSON, config.WORKSPACE_DIR, console.log, options[ 'npm-update' ] );
 	} );
 
-	gulp.task( 'dev-status', () => {
+	gulp.task( 'status', () => {
 		statusTask( ckeditor5Path, packageJSON, config.WORKSPACE_DIR, console.log, console.error );
 	} );
 
-	gulp.task( 'dev-relink', () => {
+	gulp.task( 'relink', () => {
 		relinkTask( ckeditor5Path, packageJSON, config.WORKSPACE_DIR, console.log, console.error );
 	} );
 
-	gulp.task( 'dev-install', () => {
+	gulp.task( 'install', () => {
 		const options = minimist( process.argv.slice( 2 ), {
 			string: [ 'package' ],
 			default: {
