+++ dev/tests/dev/init.js
@@ -11,7 +11,7 @@ const sinon = require( 'sinon' );
 const tools = require( '../../tasks/dev/utils/tools' );
 
 describe( 'dev-init', () => {
-	const initTask = require( '../../tasks/dev/tasks/dev-init' );
+	const initTask = require( '../../tasks/dev/tasks/init' );
 	const ckeditor5Path = 'path/to/ckeditor5';
 	const workspaceRoot = '..';
 	const emptyFn = () => {};
