+++ dev/tests/dev/install.js
@@ -11,7 +11,7 @@ const chai = require( 'chai' );
 const sinon = require( 'sinon' );
 const git = require( '../../tasks/dev/utils/git' );
 const tools = require( '../../tasks/dev/utils/tools' );
-const installTask = require( '../../tasks/dev/tasks/dev-install' );
+const installTask = require( '../../tasks/dev/tasks/install' );
 const expect = chai.expect;
 const path = require( 'path' );
