+++ dev/tests/dev/relink.js
@@ -9,21 +9,21 @@
 
 const sinon = require( 'sinon' );
 const tools = require( '../../tasks/dev/utils/tools' );
-const git = require( '../../tasks/dev/utils/git' );
 const path = require( 'path' );
 
-describe( 'dev-status', () => {
-	const statusTask = require( '../../tasks/dev/tasks/dev-status' );
+describe( 'dev-relink', () => {
+	const task = require( '../../tasks/dev/tasks/relink' );
 	const ckeditor5Path = 'path/to/ckeditor5';
+	const modulesPath = path.join( ckeditor5Path, 'node_modules' );
 	const workspaceRoot = '..';
 	const workspaceAbsolutePath = path.join( ckeditor5Path, workspaceRoot );
 	const emptyFn = () => {};
 
-	it( 'should show status of dev repositories', () => {
+	it( 'should link dev repositories', () => {
 		const dirs = [ 'ckeditor5-core', 'ckeditor5-devtest' ];
 		const getDependenciesSpy = sinon.spy( tools, 'getCKEditorDependencies' );
 		const getDirectoriesStub = sinon.stub( tools, 'getCKE5Directories' ).returns( dirs );
-		const statusStub = sinon.stub( git, 'getStatus' );
+		const linkStub = sinon.stub( tools, 'linkDirectories' );
 		const json = {
 			dependencies: {
 				'ckeditor5-core': 'ckeditor/ckeditor5-core',
@@ -32,40 +32,40 @@ describe( 'dev-status', () => {
 			}
 		};
 
-		statusTask( ckeditor5Path, json, workspaceRoot, emptyFn, emptyFn );
+		task( ckeditor5Path, json, workspaceRoot, emptyFn, emptyFn );
 
 		getDependenciesSpy.restore();
 		getDirectoriesStub.restore();
-		statusStub.restore();
+		linkStub.restore();
 
-		sinon.assert.calledTwice( statusStub );
-		sinon.assert.calledWithExactly( statusStub.firstCall, path.join( workspaceAbsolutePath, dirs[ 0 ] ) );
-		sinon.assert.calledWithExactly( statusStub.secondCall, path.join( workspaceAbsolutePath, dirs[ 1 ] ) );
+		sinon.assert.calledTwice( linkStub );
+		sinon.assert.calledWithExactly( linkStub.firstCall, path.join( workspaceAbsolutePath, dirs[ 0 ] ), path.join( modulesPath, dirs[ 0 ] ) );
+		sinon.assert.calledWithExactly( linkStub.secondCall, path.join( workspaceAbsolutePath, dirs[ 1 ] ), path.join( modulesPath, dirs[ 1 ] ) );
 	} );
 
-	it( 'should not get status when no dependencies are found', () => {
+	it( 'should not link when no dependencies are found', () => {
 		const getDependenciesSpy = sinon.spy( tools, 'getCKEditorDependencies' );
 		const getDirectoriesStub = sinon.stub( tools, 'getCKE5Directories' );
-		const statusStub = sinon.stub( git, 'getStatus' );
+		const linkStub = sinon.stub( tools, 'linkDirectories' );
 		const json = {
 			dependencies: {
 				'other-plugin': '1.2.3'
 			}
 		};
 
-		statusTask( ckeditor5Path, json, workspaceRoot, emptyFn, emptyFn );
+		task( ckeditor5Path, json, workspaceRoot, emptyFn, emptyFn );
 
 		getDependenciesSpy.restore();
 		getDirectoriesStub.restore();
-		statusStub.restore();
+		linkStub.restore();
 
-		sinon.assert.notCalled( statusStub );
+		sinon.assert.notCalled( linkStub );
 	} );
 
-	it( 'should not get status when no plugins in dev mode', () => {
+	it( 'should not link when no plugins in dev mode', () => {
 		const getDependenciesSpy = sinon.spy( tools, 'getCKEditorDependencies' );
 		const getDirectoriesStub = sinon.stub( tools, 'getCKE5Directories' ).returns( [] );
-		const statusStub = sinon.stub( git, 'getStatus' );
+		const linkStub = sinon.stub( tools, 'linkDirectories' );
 		const json = {
 			dependencies: {
 				'ckeditor5-devtest': 'ckeditor/ckeditor5-devtest#new-branch',
@@ -73,21 +73,21 @@ describe( 'dev-status', () => {
 			}
 		};
 
-		statusTask( ckeditor5Path, json, workspaceRoot, emptyFn, emptyFn );
+		task( ckeditor5Path, json, workspaceRoot, emptyFn, emptyFn );
 
 		getDependenciesSpy.restore();
 		getDirectoriesStub.restore();
-		statusStub.restore();
+		linkStub.restore();
 
-		sinon.assert.notCalled( statusStub );
+		sinon.assert.notCalled( linkStub );
 	} );
 
-	it( 'should write error message when getStatus is unsuccessful', () => {
+	it( 'should write error message when linking is unsuccessful', () => {
 		const dirs = [ 'ckeditor5-core' ];
 		const getDependenciesSpy = sinon.spy( tools, 'getCKEditorDependencies' );
 		const getDirectoriesStub = sinon.stub( tools, 'getCKE5Directories' ).returns( dirs );
 		const error = new Error( 'Error message.' );
-		const statusStub = sinon.stub( git, 'getStatus' ).throws( error );
+		const linkStub = sinon.stub( tools, 'linkDirectories' ).throws( error );
 		const json = {
 			dependencies: {
 				'ckeditor5-core': 'ckeditor/ckeditor5-core',
@@ -97,13 +97,13 @@ describe( 'dev-status', () => {
 		};
 		const writeErrorSpy = sinon.spy();
 
-		statusTask( ckeditor5Path, json, workspaceRoot, emptyFn, writeErrorSpy );
+		task( ckeditor5Path, json, workspaceRoot, emptyFn, writeErrorSpy );
 
 		getDependenciesSpy.restore();
 		getDirectoriesStub.restore();
-		statusStub.restore();
+		linkStub.restore();
 
-		sinon.assert.calledOnce( statusStub );
+		sinon.assert.calledOnce( linkStub );
 		sinon.assert.calledOnce( writeErrorSpy );
 		sinon.assert.calledWithExactly( writeErrorSpy, error );
 	} );
