+++ dev/tests/dev/status.js
@@ -9,21 +9,21 @@
 
 const sinon = require( 'sinon' );
 const tools = require( '../../tasks/dev/utils/tools' );
+const git = require( '../../tasks/dev/utils/git' );
 const path = require( 'path' );
 
-describe( 'dev-relink', () => {
-	const task = require( '../../tasks/dev/tasks/dev-relink' );
+describe( 'dev-status', () => {
+	const statusTask = require( '../../tasks/dev/tasks/status' );
 	const ckeditor5Path = 'path/to/ckeditor5';
-	const modulesPath = path.join( ckeditor5Path, 'node_modules' );
 	const workspaceRoot = '..';
 	const workspaceAbsolutePath = path.join( ckeditor5Path, workspaceRoot );
 	const emptyFn = () => {};
 
-	it( 'should link dev repositories', () => {
+	it( 'should show status of dev repositories', () => {
 		const dirs = [ 'ckeditor5-core', 'ckeditor5-devtest' ];
 		const getDependenciesSpy = sinon.spy( tools, 'getCKEditorDependencies' );
 		const getDirectoriesStub = sinon.stub( tools, 'getCKE5Directories' ).returns( dirs );
-		const linkStub = sinon.stub( tools, 'linkDirectories' );
+		const statusStub = sinon.stub( git, 'getStatus' );
 		const json = {
 			dependencies: {
 				'ckeditor5-core': 'ckeditor/ckeditor5-core',
@@ -32,40 +32,40 @@ describe( 'dev-relink', () => {
 			}
 		};
 
-		task( ckeditor5Path, json, workspaceRoot, emptyFn, emptyFn );
+		statusTask( ckeditor5Path, json, workspaceRoot, emptyFn, emptyFn );
 
 		getDependenciesSpy.restore();
 		getDirectoriesStub.restore();
-		linkStub.restore();
+		statusStub.restore();
 
-		sinon.assert.calledTwice( linkStub );
-		sinon.assert.calledWithExactly( linkStub.firstCall, path.join( workspaceAbsolutePath, dirs[ 0 ] ), path.join( modulesPath, dirs[ 0 ] ) );
-		sinon.assert.calledWithExactly( linkStub.secondCall, path.join( workspaceAbsolutePath, dirs[ 1 ] ), path.join( modulesPath, dirs[ 1 ] ) );
+		sinon.assert.calledTwice( statusStub );
+		sinon.assert.calledWithExactly( statusStub.firstCall, path.join( workspaceAbsolutePath, dirs[ 0 ] ) );
+		sinon.assert.calledWithExactly( statusStub.secondCall, path.join( workspaceAbsolutePath, dirs[ 1 ] ) );
 	} );
 
-	it( 'should not link when no dependencies are found', () => {
+	it( 'should not get status when no dependencies are found', () => {
 		const getDependenciesSpy = sinon.spy( tools, 'getCKEditorDependencies' );
 		const getDirectoriesStub = sinon.stub( tools, 'getCKE5Directories' );
-		const linkStub = sinon.stub( tools, 'linkDirectories' );
+		const statusStub = sinon.stub( git, 'getStatus' );
 		const json = {
 			dependencies: {
 				'other-plugin': '1.2.3'
 			}
 		};
 
-		task( ckeditor5Path, json, workspaceRoot, emptyFn, emptyFn );
+		statusTask( ckeditor5Path, json, workspaceRoot, emptyFn, emptyFn );
 
 		getDependenciesSpy.restore();
 		getDirectoriesStub.restore();
-		linkStub.restore();
+		statusStub.restore();
 
-		sinon.assert.notCalled( linkStub );
+		sinon.assert.notCalled( statusStub );
 	} );
 
-	it( 'should not link when no plugins in dev mode', () => {
+	it( 'should not get status when no plugins in dev mode', () => {
 		const getDependenciesSpy = sinon.spy( tools, 'getCKEditorDependencies' );
 		const getDirectoriesStub = sinon.stub( tools, 'getCKE5Directories' ).returns( [] );
-		const linkStub = sinon.stub( tools, 'linkDirectories' );
+		const statusStub = sinon.stub( git, 'getStatus' );
 		const json = {
 			dependencies: {
 				'ckeditor5-devtest': 'ckeditor/ckeditor5-devtest#new-branch',
@@ -73,21 +73,21 @@ describe( 'dev-relink', () => {
 			}
 		};
 
-		task( ckeditor5Path, json, workspaceRoot, emptyFn, emptyFn );
+		statusTask( ckeditor5Path, json, workspaceRoot, emptyFn, emptyFn );
 
 		getDependenciesSpy.restore();
 		getDirectoriesStub.restore();
-		linkStub.restore();
+		statusStub.restore();
 
-		sinon.assert.notCalled( linkStub );
+		sinon.assert.notCalled( statusStub );
 	} );
 
-	it( 'should write error message when linking is unsuccessful', () => {
+	it( 'should write error message when getStatus is unsuccessful', () => {
 		const dirs = [ 'ckeditor5-core' ];
 		const getDependenciesSpy = sinon.spy( tools, 'getCKEditorDependencies' );
 		const getDirectoriesStub = sinon.stub( tools, 'getCKE5Directories' ).returns( dirs );
 		const error = new Error( 'Error message.' );
-		const linkStub = sinon.stub( tools, 'linkDirectories' ).throws( error );
+		const statusStub = sinon.stub( git, 'getStatus' ).throws( error );
 		const json = {
 			dependencies: {
 				'ckeditor5-core': 'ckeditor/ckeditor5-core',
@@ -97,13 +97,13 @@ describe( 'dev-relink', () => {
 		};
 		const writeErrorSpy = sinon.spy();
 
-		task( ckeditor5Path, json, workspaceRoot, emptyFn, writeErrorSpy );
+		statusTask( ckeditor5Path, json, workspaceRoot, emptyFn, writeErrorSpy );
 
 		getDependenciesSpy.restore();
 		getDirectoriesStub.restore();
-		linkStub.restore();
+		statusStub.restore();
 
-		sinon.assert.calledOnce( linkStub );
+		sinon.assert.calledOnce( statusStub );
 		sinon.assert.calledOnce( writeErrorSpy );
 		sinon.assert.calledWithExactly( writeErrorSpy, error );
 	} );
