+++ dev/tests/dev/update.js
@@ -13,7 +13,7 @@ const git = require( '../../tasks/dev/utils/git' );
 const path = require( 'path' );
 
 describe( 'dev-update', () => {
-	const updateTask = require( '../../tasks/dev/tasks/dev-update' );
+	const updateTask = require( '../../tasks/dev/tasks/update' );
 	const ckeditor5Path = 'path/to/ckeditor5';
 	const workspaceRoot = '..';
 	const workspaceAbsolutePath = path.join( ckeditor5Path, workspaceRoot );
